requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anthropic==0.25.8
anyio==4.3.0
attrs==23.2.0
azure-core==1.30.1
azure-storage-blob==12.20.0
blinker==1.8.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
crewai==0.22.5
cryptography==42.0.7
dataclasses-json==0.6.5
deepgram-sdk==3.2.7
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.9.0
docstring-parser==0.15
faiss-cpu==1.8.0
filelock==3.14.0
Flask==3.0.3
Flask-Cors==4.0.1
frozenlist==1.4.1
fsspec==2024.3.1
geographiclib==2.0
geopy==2.4.1
googleapis-common-protos==1.63.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.0
idna==3.7
importlib-metadata==7.0.0
instructor==0.5.2
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.17
langchain-anthropic==0.1.11
langchain-community==0.0.37
langchain-core==0.1.52
langchain-openai==0.0.5
langchain-text-splitters==0.0.1
langsmith==0.1.55
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.26.0
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-http==1.24.0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
orjson==3.10.3
packaging==23.2
protobuf==4.25.3
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pydub==0.25.1
Pygments==2.18.0
PyPDF2==3.0.1
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rich==13.7.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
tenacity==8.3.0
tiktoken==0.5.2
tokenizers==0.19.1
tqdm==4.66.4
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.11.0
urllib3==2.2.1
verboselogs==1.7
websockets==12.0
Werkzeug==3.0.3
wrapt==1.16.0
yarl==1.9.4
zipp==3.18.1
